requirements.txt
asgiref==3.3.4
astroid==2.5.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.2
feedparser==6.0.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
pylint==2.7.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
waitress==2.0.0
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
